refactor : refactor dashboard layout
