+++ apps/seller/src/app/[locale]/_components/home.tsx
