+++ apps/seller/src/app/[locale]/page.tsx
@@ -1,5 +1,5 @@
-import { DashboardHome } from "@/components/dashboard-home";
-import { DashboardLayout } from "@/components/dashboard-layout";
+import { DashboardLayout } from "./_components";
+import { DashboardHome } from "./_components/home";
 
 export default function Home() {
   return (
