+++ packages/ui/src/index.ts
@@ -1,3 +1,4 @@
 // Export all UI components
 export * from "./components";
 export * from "./lib/utils";
+export * from "./hooks/use-mobile";
